+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - New program:      Conversion: OceanTidesDTU2GriddedData.
 - New class:        PlotDegreeAmplitudes: degreeAmplitudesSimple.
 - New option:       GnssAntennaNormalsConstraint: gnssType selection for TEC constraint.
+- New option:       PlotAxisLabeled: majorTickSpacing, minorTickSpacing, gridLineSpacing.
 - File format:      TideGeneratingPotential includes now degree 3 tides.
 - Bugfix:           GUI: fixed Ctrl+Shift+Up/Down for variables.
 - Bugfix:           slrParametrizationRangeBiasStationSatellite: Fix station index.